# openapi-appengine.yaml
# [START swagger]
swagger: "2.0"
info:
  description: "A simple Google Cloud Endpoints API - Connectivity with BigData."
  title: "People Mangement system"
  version: "1.0.0"
host: "captain-007.appspot.com"
# [END swagger]
consumes:
- "application/json"
produces:
- "application/json"
schemes:
- "https"
paths:
  "/peopleDetails":
    post:
      description: "Submit people Details."
      operationId: "postpeopledetail"
      produces:
      - "application/json"
      responses:
        200:
          description: "Data Posted"
      parameters:
      - description: "People ID"
        in: query
        name: Id
        type: integer
        required: true
      - description: "People Name"
        in: query
        name: Name
        type: string
        required: true
      - description: "People Age"
        in: query
        name: Age
        type: integer
        required: true
      - description: "People Blood_Group"
        in: query
        name: Blood_Group
        type: string
        required: true
      - description: "People Gender"
        in: query
        name: Gender
        type: string
        required: true
      - description: "People Department"
        in: query
        name: Department
        type: string
        required: true
      - description: "People DOB"
        in: query
        name: DOB
        type: string
        required: true
      - description: "People Marital_Status"
        in: query
        name: Marital_Status
        type: string
        required: true
    get:
      description: "Returns the people details"
      operationId: "getpeopledetail"
      produces:
      - "application/json"
      responses:
        200:
          description: "People info."
          schema:
            $ref: "#/definitions/PeopleDetails"
      parameters:
      - description: "People ID"
        in: query
        name: Id
        type: integer
        required: true
definitions:
  PeopleDetails:
    properties:
      Id:
        type: "integer"
        format: "int32"
        example: "5"
      Name:
        type: "string"
        format: "string"
        example: "BoB"
      Age:
        type: "integer"
        format: "int32"
        minimum: 1
        maximum: 100
        example: "99"
      Blood_Group:
        type: "string"
        format: "string"
        example: "A+"
      Gender:
        type: "string"
        format: "string"
        example: "M or F or Others"
      Department:
        type: "string"
        format: "string"
        example: "IT"
      DOB:
        type: "string"
        format: "string"
        example: "DD-MM-YYYY"
      Marital_Status:
        type: "string"
        format: "string"
        example: "yes or No"